Create binarysearch.py
Write a Python function findIntersection(L1, L2) that accepts two integer lists L1 and L2
and return a list that contains elements that are common to both lists. Try to write a solution that
runs in O(nlogn) time.
L1 contains all distinct integers and L2 contains all distinct integers, but there can be many
elements common between L1 and L2 .
List L1 intersection L2 contains all elements that are common to L1 and L2 . The elements in
the returned list can be in any order.