Create mergesort.py
You have a deck of shuffled cards ranging from 0 to 100,000,000. There are 2 sub-ordinate below
you and two subordinates below them and it goes on.
The job of the sub-ordinate is to split the deck of cards that they received and give it to two
sub-ordinate of them. If they receive a deck of cards from their subordinates, they merge it
in an ascending order and give it their higher level.
If a subordinate received only two card, then he/she himself/herself arrange in ascending
order give it back that to the superior.
If a subordinate received only one card, then he/she will give back that to the superior.
Your task is to find how many people (including you) are required to sort the cards and print the
sorted deck of cards and number of people required as a tuple
input: [194, 69, 103, 150, 151, 44, 103, 98]
output: ([44, 69, 98, 103, 103, 150, 151, 194], 7)