Create findingpair.py
Given a list L of random numbers and another number pairSum , find whether there exists two
numbers in the list such that their sum is equal to pairSum . Complete the below Python function
findPair(L, pairSum) that performs this operation. Try to write a solution which is or o(nlogn)
better.
Hint: Try to sort the list first.
For example consider the below list. We need to find if there exists any pair whose sum is equal to
21. 11+10 = 21. So the function should return True.
For the same list if we want to find if there exist any pair whose sum is equal to 2. Clearly there is
no such pair so the function should return False.
input: 3825 29286 12635 73450 69276 1052 67615 27144 32660
output: 3825