Create queue.py
A queue is a data structure in which whatever comes first will go out first. It follows the FIFO
(First-In-First-Out) policy. In Queue, the insertion is done from one end known as the rear end of
the queue, whereas the deletion is done from another end known as the front end of the queue.
For a given class Queue, create two methods:
Enqueue(data): that accepts an integer data and inserts it into the queue in the last
position.
Dequeue(): if queue is empty, returns None . Otherwise, delete one element of the queue
from the first position and returns deleted element value.
Note- Use given linked list structure for implementation.
input:
l1->[1,2,3,4,5,6,7]
l2->3
output: 
1
2
3
4,5,6,7